requirements.txt
Flask
requests
pygame
python-dotenv
SpeechRecognition